third_party/WebKit/LayoutTests/fast/text/first-letter-bad-line-boxes-crash-expected.txt

CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
CONSOLE WARNING: We don't execute document.execCommand() this time, because it is called recursively.
CONSOLE WARNING: We don't execute document.execCommand() this time, because it is called recursively.
\\\\\\m{VVVVVVVVVVVVVVV<};<WWWWW

 Te\\\\\\m{VVVVVVVVVVVVVVV<};<WWWWW
xt
